+++ src/slskd/Messaging/MessagingDbContext.cs
@@ -45,6 +45,9 @@ namespace slskd.Messaging
             modelBuilder.Entity<PrivateMessage>().HasKey(e => new { e.Username, e.Id, e.Timestamp });
             modelBuilder.Entity<PrivateMessage>().HasIndex(e => e.Username);
             modelBuilder.Entity<PrivateMessage>()
+                .HasIndex(e => new { e.Username, e.Timestamp })
+                .HasDatabaseName("IDX_PrivateMessages_Username_Timestamp");
+            modelBuilder.Entity<PrivateMessage>()
                 .HasIndex(e => new { e.IsAcknowledged, e.Username })
                 .HasDatabaseName("IDX_PrivateMessages_IsAcknowledged");